Moved Logging.Core to Internal.Logging.Core to hide functions
