+++ api/hs-opentelemetry-api.cabal
@@ -36,6 +36,7 @@ library
       OpenTelemetry.Contrib.SpanTraversals
       OpenTelemetry.Exporter
       OpenTelemetry.Internal.Common.Types
+      OpenTelemetry.Internal.Logging.Core
       OpenTelemetry.Internal.Logging.Types
       OpenTelemetry.Internal.Trace.Id
       OpenTelemetry.LogAttributes
